# refactor(holochain_state): optimize chain head query