--- a/crates/holochain/CHANGELOG.md
+++ b/crates/holochain/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 - Change `hcterm` to use the admin and app websocket clients from `holochain_client` internally
+- Reinstate indexes on `DhtOp` tables. With the latest migration script, the indexes were not carried over. #4970
+- Optimize `ChainHeadQuery` for performance. A flame graph analysis revealed that a significant proportion of the CPU time was spent on this query in a test with a high number of entry creates and reads. The query now runs about 30 % faster. #4971
 
 ## 0.6.0-dev.3