### Summary

The `ChainHeadQuery` is run every time an action is created in the source chain. A flamegraph uncovered that it was taking up a significant portion of the CPU time. This is a proposal to optimize it.

Here's a test run that exercises this query heavily, creating 5000 entries and reading them back.

Query before
```sh
running 1 test
The last 100 calls took on average 5 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 10 ms
The last 100 calls took on average 11 ms
The last 100 calls took on average 11 ms
The last 100 calls took on average 11 ms
The last 100 calls took on average 12 ms
The last 100 calls took on average 12 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 15 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 18 ms
The last 100 calls took on average 18 ms
The last 100 calls took on average 19 ms
The last 100 calls took on average 20 ms
The last 100 calls took on average 21 ms
The last 100 calls took on average 21 ms
The last 100 calls took on average 21 ms
The last 100 calls took on average 23 ms
The last 100 calls took on average 25 ms
The last 100 calls took on average 25 ms
The last 100 calls took on average 25 ms
The last 100 calls took on average 25 ms
The last 100 calls took on average 28 ms
The last 100 calls took on average 27 ms
The last 100 calls took on average 28 ms
The last 100 calls took on average 28 ms
The last 100 calls took on average 28 ms
The last 100 calls took on average 28 ms
The last 100 calls took on average 32 ms
The last 100 calls took on average 32 ms
The last 100 calls took on average 32 ms
The last 100 calls took on average 34 ms
The last 100 calls took on average 33 ms
The last 100 calls took on average 35 ms
The last 100 calls took on average 37 ms
Made 5000 calls in total over 152.306901917s.
```

Query after
```sh
running 1 test
The last 100 calls took on average 4 ms
The last 100 calls took on average 4 ms
The last 100 calls took on average 4 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 5 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 6 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 7 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 8 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 9 ms
The last 100 calls took on average 10 ms
The last 100 calls took on average 10 ms
The last 100 calls took on average 10 ms
The last 100 calls took on average 11 ms
The last 100 calls took on average 11 ms
The last 100 calls took on average 12 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 13 ms
The last 100 calls took on average 14 ms
The last 100 calls took on average 14 ms
The last 100 calls took on average 14 ms
The last 100 calls took on average 14 ms
The last 100 calls took on average 14 ms
The last 100 calls took on average 15 ms
The last 100 calls took on average 15 ms
The last 100 calls took on average 15 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 16 ms
The last 100 calls took on average 17 ms
The last 100 calls took on average 17 ms
The last 100 calls took on average 17 ms
The last 100 calls took on average 17 ms
The last 100 calls took on average 18 ms
The last 100 calls took on average 19 ms
Made 5000 calls in total over 115.961880667s.
```

see also #4079 

### TODO:
- [ ] CHANGELOGs updated with appropriate info
- [ ] All code changes are reflected in docs, including module-level docs